+++ resources/views/administrador/tramite/configuracion_tramite/tipo_tramite.blade.php
@@ -5,9 +5,12 @@
 <div class="card">
     <div class="card-header d-flex justify-content-between align-items-center">
         <h5 class="mb-0">:::::::: TIPOS DE TRAMITE :::::::: </h5>
-        <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#modal_tipo_tramite_nuevo">
-            <i class="ti ti-plus me-sm-1"></i> <span class="d-none d-sm-inline-block">Nuevo</span>
-        </button>
+        @can('tipos_tramite_submenu_nuevo')
+            <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#modal_tipo_tramite_nuevo">
+                <i class="ti ti-plus me-sm-1"></i> <span class="d-none d-sm-inline-block">Nuevo</span>
+            </button>
+        @endcan
+
     </div>
     <div class="table-responsive text-nowrap p-4">
         <table class="table table-hover" id="tabla_tipo_tramite" style="width: 100%">
@@ -138,7 +141,7 @@
                         className: 'table-td',
                         render: function(data, type, row, meta) {
                             return `
-                            @can('genero_estado')
+                            @can('tipos_tramite_submenu_estado')
                                 <label class="switch switch-primary">
                                     <input onclick="estado_tipo_tramite('${row.id}')" type="checkbox" class="switch-input" ${row.estado === true || row.estado === 1 ? 'checked' : ''} />
                                     <span class="switch-toggle-slider">
@@ -160,14 +163,17 @@
                         render: function(data, type, row, meta) {
                             return `
                                 <div class="d-inline-block tex-nowrap">
-                                    <button type="button" onclick="editar_tipo_tramite('${row.id}')" class="btn btn-icon rounded-pill btn-warning" data-toggle="tooltip" data-placement="top" title="EDITAR">
-                                        <i class="ti ti-edit" ></i>
-                                    </button>
+                                    @can('tipos_tramite_submenu_editar')
+                                        <button type="button" onclick="editar_tipo_tramite('${row.id}')" class="btn btn-icon rounded-pill btn-warning" data-toggle="tooltip" data-placement="top" title="EDITAR">
+                                            <i class="ti ti-edit" ></i>
+                                        </button>
+                                    @endcan
 
-                                    <button type="button" onclick="eliminar_tipo_tramite('${row.id}')" class="btn btn-icon rounded-pill btn-danger" data-toggle="tooltip" data-placement="top" title="ELIMINAR">
-                                        <i class="ti ti-trash" ></i>
-                                    </button>
-
+                                    @can('tipos_tramite_submenu_eliminar')
+                                        <button type="button" onclick="eliminar_tipo_tramite('${row.id}')" class="btn btn-icon rounded-pill btn-danger" data-toggle="tooltip" data-placement="top" title="ELIMINAR">
+                                            <i class="ti ti-trash" ></i>
+                                        </button>
+                                    @endcan
                                 </div>
                             `;
                         }
